% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_select.R
\name{age_select}
\alias{age_select}
\title{age_select}
\usage{
age_select(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
A list with variable names as the elements. Each element is itself a
    list with species as the elements. Each speices element is a list with
    the age range for a locale as the elements as a length 2 vector with the
    lower limit as the first value and the upper limit as the second value.
}
\description{
Age ranges for selection based on age.
}
\details{
None
}
\examples{
age_select() ## Show the entire list

    ## Create the age selection but change the lower limit for Hg of
    ## herrings at Landsort to 1.
    agesel <- age_select()
    agesel$HG$CLUP$LAND[1] <- 1
}
\author{
Erik Lampa
}
