% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkvar.R
\name{checkvar}
\alias{checkvar}
\title{checkvar}
\usage{
checkvar(mocis, var, genus)
}
\arguments{
\item{mocis}{Output from \code{mocis}}

\item{var}{A variable}

\item{genus}{Species}
}
\value{
1 if the variable exists, 0 otherwise
}
\description{
Checks if a variable exists before generating tables
}
\details{
For internal use only
}
\author{
Erik Lampa
}
