% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{convert}
\usage{
convert(data, inds, what, genus, srt, loc, co_met, co_dde, type)
}
\arguments{
\item{data}{A data frame}

\item{inds}{Variable name(s) containing the contaminants}

\item{what}{}

\item{genus}{Species}

\item{srt}{Class of contaminants, taken from the file name}

\item{loc}{Locale}

\item{co_met}{A list with conversion rules for metals, see \code{convert_met}}

\item{co_dde}{A list with conversion rules for DDE, see \code{convert_dde}}

\item{type}{Type of data (marine or limnic), Taken from the file name}
}
\value{
The converted variable
}
\description{
Converts the contaminant variables to account for lab changes
}
\details{
For metals and DDE. For internal use only.
}
\author{
Erik Lampa
}
