% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dde.R
\name{convert_dde}
\alias{convert_dde}
\title{convert_dde()}
\usage{
convert_dde(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
A list with conversion rules
}
\description{
Conversion rules for DDE
}
\details{
For use in \code{convert}. Produces a list with species names as the
    elements. Each element is itself a list with values for the locales.
}
\examples{
## Change the value for Herring at Landsort to 0.5
    conv_dde <- convert_dde()
    conv_dde$CLUP$LAND <- 0.5
}
\author{
Erik Lampa
}
