% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pcb.R
\name{convert_pcb}
\alias{convert_pcb}
\title{convert_pcb}
\usage{
convert_pcb(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
A data frame with Species, Locales, r1 and r2 values
}
\description{
Sets conversion rules for SUMPCB
}
\details{
None
}
\author{
Erik Lampa
}
