% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_out.R
\name{detect_out}
\alias{detect_out}
\title{detect_out}
\usage{
detect_out(fit, alpha)
}
\arguments{
\item{fit}{An \code{lm} fit.}

\item{alpha}{Cutoff for declaring a point an outlier, defaults to 0.05.}
}
\value{
A vector with 0 and 1 where 1 denotes the suspected outlier.
}
\description{
Marks suspected outliers in a series
}
\details{
Uses the methodology described in \ldots to check for suspected outliers.
}
\author{
Erik Lampa
}
