% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_changep.R
\name{do_changep}
\alias{do_changep}
\title{do_changep}
\usage{
do_changep(data, y, n0)
}
\arguments{
\item{data}{A data frame.}

\item{y}{The outcome variable.}

\item{n0}{The minimal number of years on each side of the change point.}
}
\value{
A list with the change point, the value of the D statistic at the
    change point and the fitted values
}
\description{
Tests for a possible change point in the series accoring to the methodology
described in \ldots
}
\details{
None
}
\author{
Erik Lampa
}
