% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_reg.R
\name{do_reg}
\alias{do_reg}
\title{do_reg}
\usage{
do_reg(data, y, alpha, rob, b.pow, n.pow, power)
}
\arguments{
\item{data}{A data frame.}

\item{y}{The outcome variable.}

\item{alpha}{Significance lavel for the power calculations.}

\item{rob}{Set to true to perform a robust linear model fit. Very experimental!}
}
\value{
A list containing the slope esimate and confidence interval
    expressed as percentage per year, the coefficient of variation around
    the regression line, three estimates of power, the coefficient of
    determination, a p-value for the null hypothesis that the slope is 0,
    the estimated concentration of the contaminant and its confidence
    interval for the last year in the series, the fitted line, the years
    present in the data and the linear model object.
}
\description{
Fits a linear model using the log-transformed contaminant values as the
outcome and year as the indpendent variable.
}
\details{
Options include a robust linear model fit using \code{lmrob} from the
    \code{robustbase} package although all power calculations are based on
    the ordinary least squares fit. The fit is returned and all methods
    available for an \code{lm} or \code{roblm} fit can be used.
}
\author{
Erik Lampa
}
