% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_smooth.R
\name{do_smooth}
\alias{do_smooth}
\title{do_smooth}
\usage{
do_smooth(data, y, alpha, crit, rob, lmod, n.pow, power)
}
\arguments{
\item{data}{A data frame.}

\item{y}{The outcome variable.}

\item{alpha}{The significance level to be used in the test against the
linear model.}

\item{crit}{A charater, "gcv" or "aic". The criterion for choosing the
optimal span.}

\item{rob}{Set to TRUE to perform a robust LOESS. Very experimental!}

\item{lmod}{The linear model which the smooth should be compared against.}
}
\value{
A list with elements cv containing the estimated coefficient of
    variation around the smooth and the lowest detectable trend the last 10
    years, an estimate of Kendall's tau, its p-value, the coefficient of
    determination, estimated contaminant level for the last year along with
    its confidence interval, the fitted smooth curve and the span used.
}
\description{
Fit a smooth curve to the series using LOESS with automatic smoothness
selection using generalized cross validation (default) or AIC to select the
optimal span. The smooth is tested against the linear model fit using an ANOVA.
}
\details{
Defaults to a linear model if for some reason a smooth cannot be fitted.
}
\author{
Erik Lampa
}
