% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fat_adjust.R
\name{fat_adjust}
\alias{fat_adjust}
\title{fat_adjust}
\usage{
fat_adjust(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
A list with contaminants, species and possibly locales where
    fatadjustment should be made.
}
\description{
Sets the rules for fat adjustments
}
\details{
For internal use only
}
\author{
Erik Lampa
}
