% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_shape.R
\name{fortify_shape}
\alias{fortify_shape}
\title{fortify_shape}
\usage{
fortify_shape(x)
}
\arguments{
\item{x}{Output from readOGR or similar}
}
\value{
A data frame
}
\description{
Sets shape data up for plotting
}
\details{
Taken from https://rpubs.com/danielequs/199150
}
\author{
https://rpubs.com/danielequs/199150
}
