% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tab.R
\name{gen_tab}
\alias{gen_tab}
\title{gen_tab}
\usage{
gen_tab(li, vars, outpath, file, ord, stationnames, specs)
}
\arguments{
\item{li}{A list resulting from \code{mocis}}

\item{vars}{Variables which the tables should be generated from. Passed from \code{mocis}.}

\item{outpath}{Where the table(s) should be saved. Passed from \code{mocis}.}

\item{file}{The file containing the data. Passed from \code{mocis}.}

\item{ord}{Ordering of the data in the table(s), defaults to \code{statorder}.}

\item{stationnames}{Names of the monitoring stations. A list, defaults to
\code{stations}.}

\item{specs}{Species list. Defaults to species().}
}
\value{
A list with elements
\itemize{
    \item genus - specis
    \item loc - the locale
    \item n - the total number of measurements
    \item yrs - the number of years in te series
    \item years.min - earliest measurement
    \item years.max - latest measurement
    \item m - the overall geometric mean value
    \item m.lower - lower confidence limit for m
    \item m.upper - upper confidence limit for m
    \item slope - slope of the regression line expressed as \% per year
    \item ci.lower - lower confidence limit for slope
    \item ci upper - upper confidence limit for slope
    \item r2 - coefficient of determination
    \item pval - p-value from the test of the null hypothesis that the slope
    is 0.
    \item cv - coefficient of variation around the linear regression line
    \item ldt - lowest detectable trend
    \item yrq - number of years required to not miss a trend
    \item pow1 - \ldots
    \item pow2 - \ldots
    \item pow3 - \ldots
    \item tau - Kendall's tau
    \item p.tau - p-value resulting from the test of the null hypothesis
    that tau equals 0.
    \item cv.sm - coefficient of determination around the regression smooth
    \item p.smooth - p-value resulting from the test with the null
    hypothesis that the smooth explains less than or as much variance as the
    regression line
    \item min.trend - minimal detectable trend during the last 10 years
    based on the degrees of freedom used by the smooth
    \item yhat.last - estimate of the contaminant level for the last year in
    the series
    \item yhat.last.lower - lower confidence limit of the estimate
    \item yhat.last.upper - upper confidence limit of the estimate
    \item slope.10 - slope for the last 10 years, if applicable, expressed
    as \% per year
    \item ci.10.lower - lower confidence limit for the slope
    \item ci.10.upper - upper confidence limit for the slope
    \item cv.10 - coefficient of variation around the slope, if applicable
    \item ldt.10 - lowest detectable trend during the last 10 years
    \item yrq.10 - number of years required
    \item pow1.10 - \ldots
    \item pow2.10 - \ldots
    \item pow3.10 - \ldots
    \item r2.10 - coefficient of determination
    \item pval.10 - p-value resulting from the test with the null hypothesis
    that the slope the last 10 years is 0
    \item changepoint - the change point, if applicable
}
}
\description{
Generates the summary table(s) resulting from the analysis.
}
\details{
For internal use only
}
\author{
Erik Lampa
}
