% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{limits}
\alias{limits}
\title{limits}
\usage{
limits(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
A list with limits for the contaminants.
}
\description{
Sets the limits to be used in the plots
}
\details{
For internal use only
}
\author{
Erik Lampa
}
