% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lod.R
\name{lod}
\alias{lod}
\title{lod}
\usage{
lod(v, fun = NULL)
}
\arguments{
\item{v}{A varaible.}

\item{fun}{A function.}
}
\value{
A value.
}
\description{
Converts values below the LOD
}
\details{
For internal use only
}
\author{
Erik Lampa
}
