% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocis.R
\name{mocis}
\alias{mocis}
\title{mocis}
\usage{
mocis(file, vars = NULL, n0 = 4, outdir = NULL, add.vars = NULL,
  sheet = 1, nyears = 5, alpha = 0.05, crit = "gcv",
  fat_adj = fat_adjust(), generate_tables = TRUE, verbose = FALSE,
  agesel = age_select(), GRU = 1, roblm = FALSE, robsm = FALSE,
  lims = limits(), ord = statorder(), stationnames = stations(),
  specs = species(), yearvar = "YWD", subset_years = 10,
  b.pow = 0.1, power = 0.8, n.pow = 10)
}
\arguments{
\item{file}{The Excel file which holds the data}

\item{vars}{Variables that should be analyzed}

\item{n0}{For use in \code{do_changep}, the minimum number of year on each
side of the change point}

\item{outdir}{Output directory. Defaults to the current working directory in
left empty.}

\item{add.vars}{Additional variables. The function tries to guess some
sensible variables if left empty.}

\item{sheet}{The sheet containing the data in the Excel file. Deafults to 1.}

\item{nyears}{The minimum number of years that should be available for
analysis. Defaults to 5.}

\item{alpha}{The p-value threshold. Defaults to 0.05}

\item{crit}{Criterion for automatic bandwidth selection. A character, "gcv"
or "aic". Degaults to "gcv".}

\item{fat_adj}{A list with rules for fat adjustment. Defaults to \code{fat_adjust()}.}

\item{generate_tables}{Logical. Should tables be generated? Deafaults to TRUE.}

\item{verbose}{Logical. If TRUE, prints the progress.}

\item{agesel}{A list with rules for the age selection. Defaults to \code{age_select}.}

\item{GRU}{For birds. Selection based on the GRU variable. Defaults to 1.}

\item{roblm}{Logical. If TRUE, performs a robust linear model. Experimental.}

\item{robsm}{Logical. If TRUE, performs a robust smooth. Experimental.}

\item{lims}{Limits for the contaminants. Defaults to \code{limits()}.}

\item{ord}{A list of \code{data.frames} giving the ordering in the
tables. Defaults to \code{statorder()}.}

\item{stationnames}{A list with station names. Defaults to
\code{stations()}.}

\item{yearvar}{The variable containing the years. Defaults to "YWD".}

\item{subset_years}{Select a subset of years to run the analyses
on. Defaults to 10.}

\item{b.pow}{The slope, as a fraction e.g. 10% should be 0.1, to use in
power calculations. Deafults to 0.1.}

\item{power}{The desired power. Befaults to 0.8.}

\item{n.pow}{The number of years for the power calculations. Defaults to 10.}
}
\value{
A list with elements:
\itemize{
   \item \code{data} A \code{data.frame} containing the raw data.
   \item \code{aggdata} A \code{data.frame} containing the aggregated data
    (yearly geometric mean values, yearly geometric standard deviations,
    number of measurements per year, if all values were below the LOD).
   \item \code{m} Overall geometric mean value with 95% confidence interval.
   \item \code{limit} A list with elements \code{tv} the target value and
    \code{fat} the average fat percentage in the series.
   \item \code{linmod} Output from \code{do_reg}.
   \item \code{linmod10} Output from \code{do_reg} applied to the last 10
    years in the series if applicable.
   \item \code{smooth} Output from \code{do_smooth}.
   \item \code{changepoint} Output from \code{do_changep}.
   \item \code{fatadjust} Either 1 or 0 dependeing on whether fat adjustment
    was done or not.
   \item \code{stattype} A character vector of length 2 denoting which type
    of file was analyzed.
   }
}
\description{
The main analysis function
}
\details{
Main function to perform the analyses, i.e. linear model fit, smooth fit,
outlier detection and creation of tables
}
\author{
Erik Lampa
}
