% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocis2d.R
\name{mocis2d}
\alias{mocis2d}
\title{mocis2d}
\usage{
mocis2d(mocis, var, genus, year = NULL, latest = 3, domain = c(9, 27,
  54, 70), bg.col = qualitative_hcl(3, "Pastel 1", l = 90)[3],
  land.col = sequential_hcl(3, "PinkYl")[3], shapefile = NULL,
  statnames = stations(), border = TRUE, size = 5,
  col.limit = diverging_hcl(2)[2], xoff = c(1, -1), yoff = c(1, -1),
  textcoord = NULL, frac = 0.8, max.lwd = 4, cuts = 5,
  specs = species(), xpos = 0.05, ypos = 0.95, cex.text = 0.8,
  plotlabels = plotlabs(), cex.mtext = 0.8, cex.legend = 1,
  pos.legend = NULL, mtext.at = NULL, ...)
}
\arguments{
\item{mocis}{The output from \code{mocis}.}

\item{var}{The contaminant of interest.}

\item{genus}{The species of interest.}

\item{year}{Which year should be the latest?}

\item{latest}{How many years should the average value be based on? Defaults
to 3.}

\item{domain}{The section of the map to be plotted. A length four vector
consisting of longitude (west), longitude (east), latitude (smouth),
latitude (north).}

\item{bg.col}{Ocean color.}

\item{land.col}{Land color.}

\item{shapefile}{Path to a shapefile if the borders should be plotted using a specific shapefile.
Defaults to NULL.}

\item{statnames}{A list with names and coordinates of the stations. Defaults to \code{stations()}.}

\item{border}{A logical. Should country borders be plotted? Deafualts to TRUE}

\item{size}{The maximum size of the circles. Defaults to 5.}

\item{col.limit}{The color of the line around the circle when the limit is exceeded.}

\item{xoff}{Adjustment to the plot area. Deafults to adding and subtracting
1 from the longitudes.}

\item{yoff}{Adjustment to the plot area. Deafults to adding and subtracting
1 from the latitudes.}

\item{textcoord}{Not used.}

\item{frac}{Not used.}

\item{max.lwd}{The maximum line with around the circle.}

\item{cuts}{Number of values to plot in the legend. Defaults to 5.}

\item{specs}{A list with species names. Defaults to \code{species{}}.}

\item{xpos}{Not used.}

\item{ypos}{Not used.}

\item{cex.text}{Size of the margin text.}

\item{plotlabels}{A list with plot labels. Defaults to \code{plotlabs()}.}

\item{cex.mtext}{Character expansion for the margin text. Defaults to 0.8.}

\item{cex.legend}{Character expansion for the legend text. Defaults to 1}

\item{pos.legend}{A length 2 vector describing the (x, y) position of the
legend.}

\item{mtext.at}{Placement of the margin text. Defaults to par("usr")[1].}

\item{...}{Other arguments passed to \code{plot()}, e.g. cex.axis to change
the character expansion for the axis tick labels.}
}
\value{
A plot.
}
\description{
Creates a 2d map showing the average concentrations of the contaminants for
each locale.
}
\details{
Spring locales are excluded. If there are \code{cuts} or less unique values those
    values are plotted, otherwise the 5th, 25th, 50th, 75th and 95th
    percentiles are plotted. Locales where the average concentration exceeds
    the limit are plotted with a red line around the circle.
}
\author{
Erik Lampa
}
