% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ne2latlong.R
\name{ne2latlong}
\alias{ne2latlong}
\title{ne2latlong}
\usage{
ne2latlong(north, east)
}
\arguments{
\item{north}{NKOO}

\item{east}{EKOO}
}
\value{
Latitude and longitude
}
\description{
Conversion of NKOO and EKOO to lat long.
}
\details{
None.
}
\author{
Erik Lampa
}
