% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mocis.R
\name{plot_mocis}
\alias{plot_mocis}
\title{plot_mocis}
\usage{
plot_mocis(mocis, var, loc, genus, nyears = 1, col.line = NULL,
  lty.line = c(1, 2), lwd.line = 2, col.sm = NULL, lty.sm = c(1,
  2), lwd.sm = 2, pval.line = c(0.05, 0.1), pval.line10 = c(0.05,
  0.1), pval.sm = c(0.05, 0.1), pch.point = 16, cex.point = 0.8,
  col.point = "black", col.lim = NULL, pch.data = 15,
  cex.data = 0.2, col.data = "black", nobs = 5, main = NULL,
  plot.changep = FALSE, xpos = 0.05, ypos = 0.95, cex.text = 0.8,
  cex.mtext = 0.8, mtext.at = NULL, bgcol = "white", ylim = NULL,
  fatadjust = FALSE, what = c("best", "line", "smooth", "last10",
  "all"), stats = stations(), specs = species(), trafo = NULL,
  outlier = "beta", col.rect = "grey90", force = FALSE,
  plotlabels = plotlabs(), statstext = TRUE, newlimit = NA, ...)
}
\arguments{
\item{mocis}{Output from \code{mocis}.}

\item{var}{A character. The contaminant of interest.}

\item{loc}{A character. The locale of interest.}

\item{genus}{A character. The species of interest.}

\item{nyears}{An integer. The lowest number of years to plot. Defaults to 1.}

\item{col.line}{A vector of two colors. Which color is used depends on the
p-value for the linear model.}

\item{lty.line}{A vector of length 2. The line types for the line. Which
line type is used depends on the p-value for the linear model. Defaults
to `c(1, 2)`.}

\item{lwd.line}{An integer. The line width for the regression line. Defaults
to 2.}

\item{col.sm}{A vector of two colors. Which color is used depends on the
p-value for the smooth.}

\item{lty.sm}{A vector of length 2. The line types for the smooth. Which
line type is used depends on the p-value for the smooth. Defaults
to \code{c(1, 2)}.}

\item{lwd.sm}{An integer. The line width for the smooth. Defaults
to 2.}

\item{pval.line}{A vector of length 2. P-value cut-offs for when the line
should be plotted. Defaults to `c(0.05, 0.1)`. If the p-value from
the linear model fit is lower than the first, the first color is used
and if the p-value is between the first and second p-values the second
color is used.}

\item{pval.line10}{A vector of length 2. P-value cut-offs for when the line
for the last 10 years should be plotted. Defaults to `c(0.05, 0.1)`.
If the p-value from the linear model fit is lower than the first,
the first color is used and if the p-value is between the first and
second p-values the second color is used.}

\item{pval.sm}{A vector of length 2. P-value cut-offs for when the smooth
should be plotted. Defaults to `c(0.05, 0.1)`. If the p-value from
the smooth is lower than the first, the first color is used
and if the p-value is between the first and second p-values the second
color is used.}

\item{pch.point}{An integer. The point style for the yearly geometric mean
values. Defaults to 16.}

\item{cex.point}{Expansion factor for the points representing the mean
values. Defaults to 0.8.}

\item{col.point}{The color of the points.}

\item{col.lim}{A vector of length 2. The colors of the polygon describing
the limits. If \code{var} is not TCDDEQV the first color is used throughout.}

\item{pch.data}{An integer. The point style for the raw data. Defaults to a
dot (.)}

\item{cex.data}{Expansion factor for the raw data points. Defaults to 0.8.}

\item{col.data}{The color of the raw data points.}

\item{nobs}{An integer. The number of observations that should be present in
a year for confidence intervals to be plotted. Defaults to 5.}

\item{main}{Not used.}

\item{plot.changep}{A logical. Should the change point be plotted (if applicable)?}

\item{xpos}{A value between 0 and 1. The position of the legend along the
x-axis in Normalized Parent Coordinates. Defaults to 0.05.}

\item{ypos}{A value between 0 and 1. The position of the legend along the
y-axis in Normalized Parent Coordinates. Defaults to 0.95.}

\item{cex.text}{Character expansion for the stats text. Defaults to 0.8.}

\item{cex.mtext}{Character expansion for the margin text. Defaults to 0.8.}

\item{mtext.at}{Position of the margin text. Defaults to par("usr")[1].}

\item{bgcol}{Not used.}

\item{ylim}{A vector of length 2. The y-axis limit.}

\item{fatadjust}{Not used.}

\item{what}{Determine what should be plotted provided the p-value(s) are in
    the specified range.  Can be one or more of:
    \code{"best"} only plot the best fit, 
    \code{"line"} only plot the line,
    \code{"smooth"} only plot the smooth,
    \code{"last10"} only plot the line for the last 10 years,
    \code{"all"} plot everything.
Defaults to "best" and "line10".}

\item{stats}{A list with station names. Defaults to \code{stations()}.}

\item{specs}{A list with species names. Defaults to \code{species()}.}

\item{trafo}{A function. Transformation for the y-axis. Experimental.}

\item{outlier}{Set to NULL to suppress marking of suspected outliers.}

\item{col.rect}{Color of the bar indicating that all measurements are below
the LOD.}

\item{force}{A logical. Force plotting even if the p-value(s) are not in the
specified range? Defaults to FALSE.}

\item{plotlabels}{A list with labels. Defaults to \code{plotlabs()}}

\item{statstext}{A logical. Controls whether to print the statistics text or
not. Defaults to TRUE.}

\item{newlimit}{Set a new limit for the plot.}

\item{...}{Other arguments to \code{plot()}.}
}
\value{
A plot.
}
\description{
Plots the data and the association between time and the contaminant
}
\details{
The default subset is the 10 last years and it is written as such in the
    help files and output regardless of the actual value used.
}
\author{
Erik Lampa
}
