% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stations.R
\name{plot_stations}
\alias{plot_stations}
\title{plot_stations}
\usage{
plot_stations(type = NULL, legend = c("right", "none"),
  bg.col = qualitative_hcl(3, "Pastel 1", l = 90)[3],
  land.col = sequential_hcl(3, "PinkYl")[3],
  col.point = qualitative_hcl(3, "Set 2")[3], cex.point = 2.5,
  stationnames = stations(), statord = statorder(), domain = c(9, 27,
  54, 70), xoff = c(1, -1), yoff = c(1, -1), legend.pos = NULL,
  marg = c(7, 4, 4, 12) + 0.1, yspace = 1.5, legend.cex = 1)
}
\arguments{
\item{type}{A character describing which stations to plot. Either
\code{"hav"} for marine stations or \code{"limn"} for
limnic stations.}

\item{legend}{Either \code{"right"} to place the legend with the station
names on the right side of the plot or \code{"none"} to suppress the
legend. Defaults to \code{"none"}.}

\item{bg.col}{Sea color.}

\item{land.col}{Land color.}

\item{col.point}{Color which the points should be filled with.}

\item{cex.point}{Expansion factor for the points.}

\item{stationnames}{A list with station names. Defaults to \code{stations()}.}

\item{statord}{The ordering of the stations. Defaults to \code{statorder}.}

\item{domain}{The section of the map to be plotted. A length four vector
consisting of longitude (west), longitude (east), latitude (smouth),
latitude (north).}

\item{xoff}{Adjustment to the plot area. Deafults to adding and subtracting
1 from the longitudes.}

\item{yoff}{Adjustment to the plot area. Deafults to adding and subtracting
1 from the latitudes.}

\item{legend.pos}{A length 2 vector of x and y coordinates. The position of
the legend if plotted.}

\item{marg}{Set the plot margins if the legend is plotted. Defaults to
\code{c(7, 4, 4, 12) + 0.1} .}

\item{yspace}{Space between the lines in the legend if plotted.}

\item{legend.cex}{Character expansion of the legend if plotted.}
}
\value{
A plot.
}
\description{
Plots a list of stations and numbers them.
}
\details{
None
}
\author{
Erik Lampa
}
