% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlabs.R
\name{plotlabs}
\alias{plotlabs}
\title{plotlabs}
\usage{
plotlabs(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
A list with plot labels
}
\description{
Plot labels for fish, bluemussels and birds
}
\details{
Nonne
}
\author{
Erik Lampa
}
