% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{pow}
\usage{
pow(b = NULL, n = NULL, a = NULL, s2 = NULL, power = NULL,
  df = NULL)
}
\arguments{
\item{b}{The trend per year as a fraction.}

\item{n}{The number of years.}

\item{a}{The confidence level.}

\item{s2}{The variance.}

\item{power}{The desired power.}

\item{df}{The degrees of freedom.}
}
\value{
Power, minimum sample size or the lowest detectable trend.
}
\description{
Function to calculate power or the minimum sample size needed to detact a
specified trend.
}
\details{
For internal use only
}
\author{
Erik Lampa
}
