% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{setup}
\usage{
setup(d, file, vars = NULL, add.vars, GRU = GRU, locs = NULL,
  yearvar = yearvar, nyears = nyears, locvar = "LOC", loc = NULL,
  agesel = agesel, co_met = convert_met(), co_pcb = convert_pcb(),
  co_dde = convert_dde(), type, srt)
}
\arguments{
\item{d}{A data.frame.}

\item{file}{The file name. Taken from \code{mocis}.}

\item{vars}{Which variables should be analyzed.}

\item{add.vars}{Additional variables to include.}

\item{GRU}{For selection among birds.}

\item{locs}{Analyze only specific locales. Not used.}

\item{yearvar}{Variable containing the years.}

\item{nyears}{How many years should be present in the series before doing
regression analysis. Taken from \code{mocis}.}

\item{locvar}{"LOC"}

\item{loc}{A vector of locales to subset the analysis on. Experimental.}

\item{agesel}{A list describing the age selection. Defaults to age_select().}

\item{co_met}{A list describing the metal conversion. Defaults to convert_met().}

\item{co_pcb}{A list describing the PCB conversion. Defaults to convert_pcb().}

\item{co_dde}{A list describing the DDE conversion. Defaults to convert_dde().}

\item{type}{Type of file.}

\item{srt}{Type of file.}
}
\value{
A list with each locale as the elements. Each locale is a list with
    the species as the elements. Each species is a list with a
    \code{data.frame} containing each contaminant as the elements.
}
\description{
Process the raw data for analysis with \code{mocis}.
}
\details{
For internal use only.
}
\author{
Erik Lampa
}
