% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations}
\alias{stations}
\title{stations}
\usage{
stations(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
A list with station names and coordinates
}
\description{
A list with station names and coordinates
}
\details{
None
}
\author{
Erik Lampa
}
