% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statorder.R
\name{statorder}
\alias{statorder}
\title{statorder}
\usage{
statorder(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
A list with \code{data.frames} describing the ordering in the tables.
}
\description{
Ordering in the tables
}
\details{
None
}
\author{
Erik Lampa
}
