% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_age.R
\name{subset_age}
\alias{subset_age}
\title{subset_age}
\usage{
subset_age(data, agesel, what, genus, where)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{agesel}{A list from age_select() or similar.}

\item{what}{A variable.}

\item{genus}{The species.}

\item{where}{The locale.}
}
\value{
Data subsetted on age.
}
\description{
Subsets the data based on the selection rules.
}
\details{
None
}
\author{
Erik Lampa
}
