% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_shape.R
\name{subset_shape}
\alias{subset_shape}
\title{subset_shape}
\usage{
subset_shape(x, domain)
}
\arguments{
\item{x}{A \code{data.frame}.}

\item{domain}{The domain.}
}
\value{
Subsetted map info.
}
\description{
Helper function.
}
\details{
Taken from https://rpubs.com/danielequs/199150
}
\author{
Erik Lampa
}
